pyproject.toml
[project]
name = "tangarin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "chainlit>=2.9.6",
    "faiss-cpu>=1.13.2",
    "greenlet>=3.3.1",
    "ipykernel>=7.2.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "psycopg>=3.3.2",
    "pypdf>=6.9.1",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.3.0",
    "sqlalchemy>=2.0.46",
    "unstructured>=0.21.5",
]
